refactor(tui): split task_screen.py into focused modules
